Saved plot of price vs. year and mileage.
While it is expected that price would correlate positively with year, and negatively with mileage, the year correlation is much stronger. At later years, this correlation appears to break down with a wide variety of prices for model years between 2013-2016. I suspect that prices below $5000 for these years may have been scraped incorrectly (there are even some points suspiciously close to 0 for all years), but the variation is large even without these. It is also interesting to note that mileage is not available for many of the later model cars, which makes sense because a consumer would probably expect this mileage is low.