Plot of the average price and mileage for each region compared to the average for all listings.
Listings from region 'nwb' (Northwest/Merrimack) appear to offer cars at a far lower mileage than the average, at a slightly higher price. In region 'gbs' (Boston/Cambridge/Brookline), average mileage of cars is high but listing price is on par with average anyway, perhaps owing to high cost of living closest to the city. In contrast, listings from 'nos' (North Shore) have cars with average mileage that is on par with the overall average but offers them at a lower price. It would be interesting to compare this data with factors such as cost of living in these regions.

In the future it would be nice to print the actual region name instead of the 3 digit craigslist code.